+++ quizzer/src/components/Quiz/QuizData.js
@@ -1,121 +1,140 @@
-const QuizData = [
-    {
-    "id": 1,
-    "category": `Math`,
-    "type": 1,
-    "Q_content": [`What is the square route of 256?`],
-    "A": "12",
-    "B": `16`,
-    "C": `14`,
-    "D": `15`,
-    "correct_answer": `16`,
-    "points": 2,
-    "options": [`12`,`14`,`15`,`16`],
-    "quiz_id": null
-    },
-    
-    {
-    "id": 2,
-    "category": `Science`,
-    "type": 1,
-    "Q_content": [`What is the chemical formula for water?`],
-    "A": `H3O`,
-    "B": `HO2`,
-    "C": `H2O2`,
-    "D": `H2O`,
-    "correct_answer": `H2O`,
-    "points": 1,
-    "options": [`H2O2`,`H30`,`HO2`, `H2O`],
-    "quiz_id": null
-    },
+import React, { useState, useEffect } from "react";
+import axios from "axios";
+import styled from "styled-components";
 
-    {
-    "id": 3,
-    "category": `English`,
-    "type": 1,
-    "Q_content": [`Which of these is a correlative conjunction function?`],
-    "A": `Although`,
-    "B": `Since`,
-    "C": `Neither/nor`,
-    "D": `After`,
-    "correct_answer": [`Neither nor`],
-    "points": 2,
-    "options": [`After`,`Neither nor`,`Since`,`Although`],
-    "quiz_id": null
-    },
+const QuizPage = styled.div`
+  text-align: center;
+`;
 
+function QuizData2() {
+  const [name] = useState('Questions');
+  const [questions, setQuestions] = useState([]);
 
-    // {
-    // "id": 4,
-    // "category": "History",
-    // "type": "multiple",
-    // "difficulty": "medium",
-    // "question": "What was the immediate trigger for the start of WWI?",
-    // "correct_answer": "Assassination of Archduke Franz Ferdinand",
-    // "incorrect_answers": "{\"Assassination of Adolf Hitler\",\"Assassination of Benito Mussolini\",\"Assassination of Joseph Stalin\"}",
-    // "points": 2
-    // },
-    // {
-    // "id": 5,
-    // "category": "Math",
-    // "type": "multiple",
-    // "difficulty": "easy",
-    // "question": "Which of these is a non-real number?",
-    // "correct_answer": "2i",
-    // "incorrect_answers": "{\"1\",\"2990\",\"-.000002\"}",
-    // "points": 1
-    // },
-    // {
-    // "id": 6,
-    // "category": "Science",
-    // "type": "multiple",
-    // "difficulty": "hard",
-    // "question": "Brass gets discolored in air because of the presence of which of the following gases in air?",
-    // "correct_answer": "Hydrogen sulphide",
-    // "incorrect_answers": "{\"Oxygen\",\"Carbon dioxide\",\"Nitrogen\"}",
-    // "points": 3
-    // },
-    // {
-    // "id": 7,
-    // "category": "English",
-    // "type": "multiple",
-    // "difficulty": "easy",
-    // "question": "Define the meaning of the idiom used in this question: He passed himself off as a noble man?",
-    // "correct_answer": "Pretending to be",
-    // "incorrect_answers": "{\"Was looked upon\",\"Was regarded as\",\"Was thought to be\"}",
-    // "points": 1
-    // },
-    // {
-    // "id": 8,
-    // "category": "History",
-    // "type": "multiple",
-    // "difficulty": "medium",
-    // "question": "What country gained its independence from the U.S. on July 4, 1946?",
-    // "correct_answer": "Philippines",
-    // "incorrect_answers": "{\"Cuba\",\"Guam\",\"Estonia\"}",
-    // "points": 2
-    // },
-    // {
-    // "id": 9,
-    // "category": "Math",
-    // "type": "multiple",
-    // "difficulty": "medium",
-    // "question": "What is the value of x + x(x^x) when x = 2?",
-    // "correct_answer": "10",
-    // "incorrect_answers": "{\"16\",\"18\",\"36\"}",
-    // "points": 2
-    // },
-    // {
-    // "id": 10,
-    // "category": "Science",
-    // "type": "multiple",
-    // "difficulty": "hard",
-    // "question": "Which of the following is a non metal that remains liquid at room temperature?",
-    // "correct_answer": "Bromine",
-    // "incorrect_answers": "{\"Phosphorous\",\"Chlorine\",\"Helium\"}",
-    // "points": 3
-    // }
-    
-];
+  //takes place instead of componentDidMount
+  useEffect(() => {
+    const fetchData = async () => {
+      const result = await axios(
+        "https://labs13-quizzer.herokuapp.com/api/quest/question"
+      );
+      //setting database data to state with hooks
+      console.log("online");
+      setQuestions(result.data);
+    };
+    fetchData();
+  }, []);
 
-export default QuizData;
+  return (
+    <QuizPage>
+      <p>{name}</p>
+      {console.log("online")}
+      {questions.map(questions => (
+        <div key={questions.id}>
+          <p>Category: {questions.category}</p>
+          <p>Type: {questions.type}</p>
+          <p>Question: {questions.Q_content}</p> 
+          <p>A: {questions.A}</p>
+          <p>B: {questions.B}</p>
+          <p>C: {questions.C}</p> 
+          <p>D: {questions.D}</p>         
+          <p>Points: {questions.points}</p>
+          <br/>
+        </div>
+      ))}
+    </QuizPage>
+  );
+}
+
+export default QuizData2;
+
+// const QuizData2 = [
+// // Quiz Q's
+//   {
+//     "id": 1,
+//     "category": `Math`,
+//     "type": 1,
+//     "Q_content": [`What is the square route of 256?`],
+//     "options": [
+//         `A: 12`,
+//         `B: 16`, 
+//         `C: 14`, 
+//         `D: 15`
+//     ],
+//     "correct_answer": `B`,
+//     "points": 2
+//   },
+
+//   {
+//     id: 2,
+//     category: "Science",
+//     type: 1,
+//     Q_content: ["What is the chemical formula for water?"],
+//     "options": [
+//         `A: H3O`,
+//         `B: HO2`,
+//         `C: H2O2`,
+//         `D: H2O`
+//     ],   
+//     correct_answer: `D`,
+//     points: 2
+//   },
+
+//   {
+//     "id": 3,
+//     "category": `English`,
+//     "type": 1,
+//     "Q_content": [`Which of these is a correlative conjunction function?`],
+//     "options": [ 
+//         `A: "Although"`,
+//         `B: "Since"`,
+//         `C: "Neither/nor"`,
+//         `D: "After"`
+//     ],
+//     "correct_answer": `C`,
+//     "points": 2
+//   },
+// // Remedial Q's
+//   {
+//     "id": 1,
+//     "category": `Math`,
+//     "type": 2,
+//     "Q_content": [`In ∆XYZ, ∠X = 55° and ∠Y = 75°. Find ∠Z?`],
+//     "options": [
+//         `A: "∠Z = 50°"`,
+//         `B: "∠Z = 150°"`,
+//         `C: "∠Z = 25°"`,
+//         `D: "∠Z = 75°"`
+//     ],
+//     "correct_answer": `A`
+//   },
+
+//   {
+//     "id": 2,
+//     "category": `Science`,
+//     "type": 2,
+//     "Q_content": [`Which of the following is used in pencils?`],
+//     "options": [
+//         `A: "Charcoal"`,
+//         `B: "Graphite"`,
+//         `C: "Phosphorous"`,
+//         `D: "Lead"`
+//     ],
+//     "correct_answer": `B`
+//   },
+
+//   {
+//     "id": 3,
+//     "category": `English`,
+//     "type": 2,
+//     "Q_content": [`Analogies: Odometer is to mileage as compass is to?`],
+//     "options": [
+//         `A: "Needle"`,
+//         `B: "Speed"`,
+//         `C: "Direction"`,
+//         `D: "Hiking"`
+//     ],
+//     "correct_answer": "C"
+//   }
+
+// ];
+
+// export default QuizData2;
